Revert "Refactor file structure"
