+++ .storybook/main.ts
@@ -3,7 +3,6 @@ import type { StorybookConfig } from "@storybook/react-webpack5";
 
 const config: StorybookConfig = {
   stories: [
-    "../src/**/*.@(mdx)",
     "../src/**/*.stories.@(js|jsx|mjs|ts|tsx)"
   ],
   addons: [
