+++ blog/2024-09-14-docusaurus-styles-in-storybook/2024-09-14-docusaurus-styles-in-storybook.mdx
@@ -6,7 +6,7 @@ tags:
     - storybook
     - formatting
 ---
-import Expandable from '@site/src/components/Expandable/Expandable';
+import Expandable from '@site/src/components/Expandable';
 
 I thought it would be a good idea to have a central location to showcase
 whatever components I add to the Docusaurus repertoire.
