+++ src/stories/components/brand/logo/Logo.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import type { Meta, StoryObj } from '@storybook/react';
-import Logo from './Logo';
+import Logo from '../../../../components/brand/Logo/Logo';
 
 const containerStyle: React.CSSProperties = {
     display: 'flex',
